Extract shutdown_otel_providers into spec/datadog/opentelemetry/spec_helper
Move the shutdown_otel_providers method to a shared OpenTelemetryHelpers
module, following the pattern at spec/datadog/di/spec_helper.rb. Each
OT spec requires it and 'include OpenTelemetryHelpers' in its top-level
describe block. The after blocks in metrics_spec.rb and logs_spec.rb
collapse from a provider.shutdown line plus a defined? + is_a? guarded
inline block into a single helper call.
