+++ spec/datadog/opentelemetry_spec.rb
@@ -1,24 +1,10 @@
 require 'spec_helper'
 require 'opentelemetry/sdk'
 require 'datadog/opentelemetry'
+require 'datadog/opentelemetry/spec_helper'
 
 RSpec.describe Datadog::OpenTelemetry do
-  # Shut down the metrics and logs providers spawned by OpenTelemetry::SDK.configure
-  # so their background threads (PeriodicMetricReader / BatchLogRecordProcessor)
-  # don't outlive the example. Each provider only has a shutdown method on the
-  # real SDK class; the Noop default returned when the corresponding SDK gem
-  # isn't loaded does not, so guard with defined? + is_a?.
-  def shutdown_otel_providers
-    if defined?(::OpenTelemetry::SDK::Metrics::MeterProvider) &&
-        ::OpenTelemetry.meter_provider.is_a?(::OpenTelemetry::SDK::Metrics::MeterProvider)
-      ::OpenTelemetry.meter_provider.shutdown
-    end
-    if defined?(::OpenTelemetry::SDK::Logs::LoggerProvider) &&
-        ::OpenTelemetry.respond_to?(:logger_provider) &&
-        ::OpenTelemetry.logger_provider.is_a?(::OpenTelemetry::SDK::Logs::LoggerProvider)
-      ::OpenTelemetry.logger_provider.shutdown
-    end
-  end
+  include OpenTelemetryHelpers
 
   context 'with Datadog TraceProvider' do
     let(:otel_tracer) { OpenTelemetry.tracer_provider.tracer('otel-tracer') }
